Fix 3 known limitations: GeneratorExit crash, Cog split, AskUser improvements
Limitation 1: SDK disconnect crash on GeneratorExit
- In send_message(), catch GeneratorExit before BaseException
- When caller breaks out of async-for loop, keep session alive
- Only real exceptions trigger disconnect; GeneratorExit just returns
- Added test: test_send_message_generator_exit_keeps_session_active

Limitation 2: Split bot.py (2284 lines) into Cog modules
- Created src/clauded/cogs/ with 7 command modules:
  - project.py: /project and /env groups (365 lines)
  - session.py: /session group (365 lines)
  - model.py: /model, /effort, /max-turns, /fallback-model, /bare (119 lines)
  - tools.py: /tools and /budget groups (154 lines)
  - agent.py: /agent group (116 lines)
  - mcp.py: /mcp group (111 lines)
  - ops.py: /cost, /health, /ratelimit, /review, /plugin, /debug,
            /notify, context menus (374 lines)
- bot.py reduced from 2284 to 767 lines (core bot logic only)
- All 21 command registrations preserved with exact parity
- SystemPromptModal re-exported from bot.py for backward compat

Limitation 3: AskUserQuestion improvements
- AskPaginatedSelectView: handles >25 options with prev/next buttons
- AskTextModal + _ModalTriggerView: free-text input via Discord Modal
  for open-ended questions (no options provided)
- InteractionHandler._ask_free_text(): renders button->modal flow
- Updated _ask_one() to route >25 options to paginated view

All 205 tests pass (204 original + 1 new GeneratorExit test).
