+++ src/clauded/cogs/__init__.py
@@ -1,0 +1,1 @@
+"""Discord command Cogs for claudeD."""
