+++ src/clauded/cogs/project.py
@@ -1,0 +1,365 @@
+"""Project management commands: /project and /env groups."""
+
+from __future__ import annotations
+
+import logging
+from pathlib import Path
+
+import discord
+from discord import app_commands
+
+from ..discord_renderer import COLOR_INFO, COLOR_TOOL_FAILURE
+from ..project_manager import ProjectManager
+
+log = logging.getLogger("clauded.bot")
+
+
+# ---------------------------------------------------------------------------
+# /project group
+# ---------------------------------------------------------------------------
+
+project_group = app_commands.Group(
+    name="project",
+    description="Manage channel ↔ project-directory bindings.",
+    default_permissions=discord.Permissions(administrator=True),
+)
+
+
+@project_group.command(name="bind", description="Bind this channel to a local directory.")
+@app_commands.describe(path="Absolute path to the project directory")
+async def project_bind(interaction: discord.Interaction, path: str) -> None:
+    log.info("/project bind path=%s channel=%s", path, interaction.channel_id)
+    from ..bot import ClaudedBot
+    bot: ClaudedBot = interaction.client  # type: ignore[assignment]
+    channel_id = interaction.channel_id
+    if channel_id is None:
+        await interaction.response.send_message(
+            "Cannot bind: no channel context.", ephemeral=True
+        )
+        return
+
+    try:
+        stored = bot.project_manager.bind(channel_id, path, guild_id=interaction.guild_id)
+    except ValueError as exc:
+        await interaction.response.send_message(f"❌ {exc}", ephemeral=True)
+        return
+
+    await interaction.response.send_message(
+        f"✅ Bound this channel to `{stored}`", ephemeral=True
+    )
+
+
+@project_group.command(name="info", description="Show this channel's current binding.")
+async def project_info(interaction: discord.Interaction) -> None:
+    log.info("/project info channel=%s", interaction.channel_id)
+    from ..bot import ClaudedBot
+    bot: ClaudedBot = interaction.client  # type: ignore[assignment]
+    channel_id = interaction.channel_id
+    if channel_id is None:
+        await interaction.response.send_message("No channel context.", ephemeral=True)
+        return
+
+    bound_path = bot.project_manager.get_project(channel_id)
+    if bound_path is None:
+        await interaction.response.send_message(
+            "This channel is not bound to a project. Use `/project bind` to set one.",
+            ephemeral=True,
+        )
+        return
+
+    lines = [f"📁 This channel is bound to `{bound_path}`"]
+    sp = bot.project_manager.get_system_prompt(channel_id)
+    if sp:
+        lines.append(f"📝 System prompt: {sp}")
+    mode = bot.project_manager.get_channel_mode(channel_id)
+    if mode != "thread":
+        lines.append(f"🔀 Channel mode: `{mode}`")
+    guild_root = bot.project_manager.get_guild_root(interaction.guild_id)
+    if interaction.guild_id and str(interaction.guild_id) in bot.project_manager._guild_roots:
+        lines.append(f"🏠 Guild root: `{guild_root}`")
+    await interaction.response.send_message("\n".join(lines), ephemeral=True)
+
+
+@project_group.command(name="unbind", description="Remove this channel's binding.")
+async def project_unbind(interaction: discord.Interaction) -> None:
+    log.info("/project unbind channel=%s", interaction.channel_id)
+    from ..bot import ClaudedBot
+    bot: ClaudedBot = interaction.client  # type: ignore[assignment]
+    channel_id = interaction.channel_id
+    if channel_id is None:
+        await interaction.response.send_message("No channel context.", ephemeral=True)
+        return
+
+    if bot.project_manager.unbind(channel_id):
+        await interaction.response.send_message(
+            "✅ Removed this channel's project binding.", ephemeral=True
+        )
+    else:
+        await interaction.response.send_message(
+            "This channel had no binding to remove.", ephemeral=True
+        )
+
+
+class SystemPromptModal(discord.ui.Modal, title="Set System Prompt"):
+    """Modal dialog for editing the channel's system prompt."""
+
+    prompt_input = discord.ui.TextInput(
+        label="System Prompt",
+        style=discord.TextStyle.paragraph,
+        placeholder="Describe the project context, coding style, etc.",
+        max_length=4000,
+        required=False,
+    )
+
+    def __init__(self, channel_id: int, project_manager: ProjectManager) -> None:
+        super().__init__()
+        self._channel_id = channel_id
+        self._pm = project_manager
+        existing = project_manager.get_system_prompt(channel_id)
+        if existing:
+            self.prompt_input.default = existing
+
+    async def on_submit(self, interaction: discord.Interaction) -> None:
+        text = self.prompt_input.value.strip()
+        if text:
+            self._pm.set_system_prompt(self._channel_id, text)
+            await interaction.response.send_message(
+                embed=discord.Embed(
+                    title="✅ System prompt updated",
+                    description=f"```\n{text[:200]}\n```",
+                    color=COLOR_INFO,
+                ),
+                ephemeral=True,
+            )
+        else:
+            self._pm.clear_system_prompt(self._channel_id)
+            await interaction.response.send_message("✅ System prompt cleared.", ephemeral=True)
+
+
+@project_group.command(name="system-prompt", description="Set system prompt for this project")
+async def project_system_prompt(interaction: discord.Interaction) -> None:
+    log.info("/project system-prompt channel=%s", interaction.channel_id)
+    from ..bot import ClaudedBot
+    bot: ClaudedBot = interaction.client  # type: ignore[assignment]
+    channel_id = interaction.channel_id
+    if channel_id is None:
+        await interaction.response.send_message("No channel context.", ephemeral=True)
+        return
+
+    if not bot.project_manager.is_bound(channel_id):
+        await interaction.response.send_message("Channel not bound. Use /project bind first.", ephemeral=True)
+        return
+
+    modal = SystemPromptModal(channel_id, bot.project_manager)
+    await interaction.response.send_modal(modal)
+
+
+@project_group.command(name="add-dir", description="Add extra directory access for Claude")
+@app_commands.describe(path="Path to directory")
+async def project_add_dir(interaction: discord.Interaction, path: str) -> None:
+    log.info("/project add-dir path=%s channel=%s", path, interaction.channel_id)
+    from ..bot import ClaudedBot
+    bot: ClaudedBot = interaction.client  # type: ignore[assignment]
+    channel_id = interaction.channel_id
+    if channel_id is None:
+        await interaction.response.send_message("No channel context.", ephemeral=True)
+        return
+    if not bot.project_manager.is_bound(channel_id):
+        await interaction.response.send_message("Channel not bound. Use /project bind first.", ephemeral=True)
+        return
+    try:
+        resolved = bot.project_manager.add_extra_dir(channel_id, path)
+    except ValueError as exc:
+        await interaction.response.send_message(f"❌ {exc}", ephemeral=True)
+        return
+    embed = discord.Embed(
+        title="📂 Extra Directory Added",
+        description=f"Added `{resolved}`",
+        color=COLOR_INFO,
+    )
+    await interaction.response.send_message(embed=embed, ephemeral=True)
+
+
+@project_group.command(name="dirs", description="List extra directories")
+async def project_dirs(interaction: discord.Interaction) -> None:
+    log.info("/project dirs channel=%s", interaction.channel_id)
+    from ..bot import ClaudedBot
+    bot: ClaudedBot = interaction.client  # type: ignore[assignment]
+    channel_id = interaction.channel_id
+    if channel_id is None:
+        await interaction.response.send_message("No channel context.", ephemeral=True)
+        return
+    dirs = bot.project_manager.get_extra_dirs(channel_id)
+    if not dirs:
+        await interaction.response.send_message("No extra directories configured.", ephemeral=True)
+        return
+    listing = "\n".join(f"• `{d}`" for d in dirs)
+    embed = discord.Embed(
+        title="📂 Extra Directories",
+        description=listing,
+        color=COLOR_INFO,
+    )
+    await interaction.response.send_message(embed=embed, ephemeral=True)
+
+
+@project_group.command(name="remove-dir", description="Remove extra directory")
+@app_commands.describe(path="Path to remove")
+async def project_remove_dir(interaction: discord.Interaction, path: str) -> None:
+    log.info("/project remove-dir path=%s channel=%s", path, interaction.channel_id)
+    from ..bot import ClaudedBot
+    bot: ClaudedBot = interaction.client  # type: ignore[assignment]
+    channel_id = interaction.channel_id
+    if channel_id is None:
+        await interaction.response.send_message("No channel context.", ephemeral=True)
+        return
+    removed = bot.project_manager.remove_extra_dir(channel_id, path)
+    if removed:
+        embed = discord.Embed(
+            title="📂 Extra Directory Removed",
+            description=f"Removed `{path}`",
+            color=COLOR_INFO,
+        )
+        await interaction.response.send_message(embed=embed, ephemeral=True)
+    else:
+        await interaction.response.send_message("Directory not found in extra dirs.", ephemeral=True)
+
+
+@project_group.command(name="set-mode", description="Set channel mode (thread or forum)")
+@app_commands.describe(mode="thread (default) or forum")
+@app_commands.choices(mode=[
+    app_commands.Choice(name="thread", value="thread"),
+    app_commands.Choice(name="forum", value="forum"),
+])
+async def project_set_mode(interaction: discord.Interaction, mode: app_commands.Choice[str]) -> None:
+    log.info("/project set-mode mode=%s channel=%s", mode.value, interaction.channel_id)
+    from ..bot import ClaudedBot
+    bot: ClaudedBot = interaction.client  # type: ignore[assignment]
+    channel_id = interaction.channel_id
+    if channel_id is None:
+        await interaction.response.send_message("No channel context.", ephemeral=True)
+        return
+    bot.project_manager.set_channel_mode(channel_id, mode.value)
+    await interaction.response.send_message(
+        f"✅ Channel mode set to `{mode.value}`", ephemeral=True
+    )
+
+
+@project_group.command(name="set-root", description="Set per-guild projects root directory")
+@app_commands.describe(path="Absolute path to the guild's projects root directory")
+async def project_set_root(interaction: discord.Interaction, path: str) -> None:
+    log.info("/project set-root path=%s guild=%s", path, interaction.guild_id)
+    from ..bot import ClaudedBot
+    bot: ClaudedBot = interaction.client  # type: ignore[assignment]
+    guild_id = interaction.guild_id
+    if guild_id is None:
+        await interaction.response.send_message(
+            "This command can only be used in a guild.", ephemeral=True
+        )
+        return
+    try:
+        resolved = bot.project_manager.set_guild_root(guild_id, path)
+    except ValueError as exc:
+        await interaction.response.send_message(f"❌ {exc}", ephemeral=True)
+        return
+    await interaction.response.send_message(
+        f"✅ Guild projects root set to `{resolved}`", ephemeral=True
+    )
+
+
+@project_group.command(name="clear-root", description="Remove per-guild projects root override")
+async def project_clear_root(interaction: discord.Interaction) -> None:
+    log.info("/project clear-root guild=%s", interaction.guild_id)
+    from ..bot import ClaudedBot
+    bot: ClaudedBot = interaction.client  # type: ignore[assignment]
+    guild_id = interaction.guild_id
+    if guild_id is None:
+        await interaction.response.send_message(
+            "This command can only be used in a guild.", ephemeral=True
+        )
+        return
+    if bot.project_manager.clear_guild_root(guild_id):
+        await interaction.response.send_message(
+            "✅ Guild projects root override removed. Using default.", ephemeral=True
+        )
+    else:
+        await interaction.response.send_message(
+            "No guild-specific root was set.", ephemeral=True
+        )
+
+
+# ---------------------------------------------------------------------------
+# /env group
+# ---------------------------------------------------------------------------
+
+env_group = app_commands.Group(
+    name="env",
+    description="Manage environment variables for Claude sessions.",
+    default_permissions=discord.Permissions(administrator=True),
+)
+
+
+@env_group.command(name="set", description="Set environment variable")
+@app_commands.describe(key="Variable name", value="Variable value")
+async def env_set(interaction: discord.Interaction, key: str, value: str) -> None:
+    from ..bot import ClaudedBot
+    bot = interaction.client
+    if not isinstance(bot, ClaudedBot):
+        await interaction.response.send_message("Bot not ready.", ephemeral=True)
+        return
+    channel_id = interaction.channel_id
+    parent_id = getattr(interaction.channel, "parent_id", None) or channel_id
+    if not bot.project_manager.is_bound(parent_id):
+        await interaction.response.send_message("Channel not bound. Use /project bind first.", ephemeral=True)
+        return
+    bot.project_manager.set_env(parent_id, key, value)
+    embed = discord.Embed(
+        title="✅ Environment Variable Set",
+        description=f"`{key}` = `{value[:100]}{'…' if len(value) > 100 else ''}`",
+        color=COLOR_INFO,
+    )
+    await interaction.response.send_message(embed=embed, ephemeral=True)
+
+
+@env_group.command(name="list", description="List environment variables")
+async def env_list(interaction: discord.Interaction) -> None:
+    from ..bot import ClaudedBot
+    bot = interaction.client
+    if not isinstance(bot, ClaudedBot):
+        await interaction.response.send_message("Bot not ready.", ephemeral=True)
+        return
+    channel_id = interaction.channel_id
+    parent_id = getattr(interaction.channel, "parent_id", None) or channel_id
+    env = bot.project_manager.get_env(parent_id)
+    if not env:
+        await interaction.response.send_message("No environment variables configured.", ephemeral=True)
+        return
+    def _mask(v: str) -> str:
+        return v[:2] + "****" if len(v) > 4 else "****"
+    listing = "\n".join(f"• `{k}` = `{_mask(v)}`" for k, v in env.items())
+    embed = discord.Embed(
+        title="🔐 Environment Variables",
+        description=listing,
+        color=COLOR_INFO,
+    )
+    await interaction.response.send_message(embed=embed, ephemeral=True)
+
+
+@env_group.command(name="remove", description="Remove environment variable")
+@app_commands.describe(key="Variable name")
+async def env_remove(interaction: discord.Interaction, key: str) -> None:
+    from ..bot import ClaudedBot
+    bot = interaction.client
+    if not isinstance(bot, ClaudedBot):
+        await interaction.response.send_message("Bot not ready.", ephemeral=True)
+        return
+    channel_id = interaction.channel_id
+    parent_id = getattr(interaction.channel, "parent_id", None) or channel_id
+    if bot.project_manager.remove_env(parent_id, key):
+        embed = discord.Embed(
+            title="✅ Environment Variable Removed",
+            description=f"Removed `{key}`",
+            color=COLOR_INFO,
+        )
+        await interaction.response.send_message(embed=embed, ephemeral=True)
+    else:
+        await interaction.response.send_message(f"Variable `{key}` not found.", ephemeral=True)
