+++ src/clauded/interaction_handler.py
@@ -11,6 +11,8 @@ tool input with an ``answers`` field — that becomes the ``updated_input`` of a
 UI rules:
 * single-select, ≤ 4 options → row of buttons
 * multi-select OR > 4 options → select menu (capped at Discord's 25 options)
+* > 25 options → paginated select menu with prev/next buttons
+* no options (open-ended question) → button that opens a text-input Modal
 
 A single ``AskUserQuestion`` invocation may carry multiple sub-questions; they
 are asked one at a time, in order. If the user fails to answer any of them
@@ -79,10 +81,10 @@ class InteractionHandler:
             multi_select = bool(q.get("multiSelect", False))
 
             if not options:
-                # No choices to render → skip silently.
-                continue
-
-            answer = await self._ask_one(q_text, header, options, multi_select)
+                # No choices — open-ended free-text question.
+                answer = await self._ask_free_text(q_text, header)
+            else:
+                answer = await self._ask_one(q_text, header, options, multi_select)
             if answer is None:
                 log.info("AskUserQuestion timed out waiting for user")
                 return None
@@ -113,6 +115,10 @@ class InteractionHandler:
         use_buttons = (not multi_select) and len(options) <= 4
         if use_buttons:
             view: _BaseAskView = AskButtonView(labels, timeout=self.timeout)
+        elif len(options) > _DISCORD_SELECT_MAX_OPTIONS:
+            view = AskPaginatedSelectView(
+                labels, descriptions, multi_select=multi_select, timeout=self.timeout
+            )
         else:
             view = AskSelectView(
                 labels, descriptions, multi_select=multi_select, timeout=self.timeout
@@ -148,6 +154,39 @@ class InteractionHandler:
         if 0 <= first < len(labels):
             return labels[first]
         return None
+
+    async def _ask_free_text(
+        self,
+        question_text: str,
+        header: str,
+    ) -> str | None:
+        """Show a button that opens a Modal for free-text input."""
+        embed = discord.Embed(
+            title=f"❓ {header}"[:256],
+            description=question_text[:_DISCORD_EMBED_DESC_MAX] or "\u200b",
+        )
+
+        view = _ModalTriggerView(
+            question=question_text,
+            header=header,
+            timeout=self.timeout,
+        )
+
+        try:
+            msg = await self.thread.send(embed=embed, view=view)
+        except discord.HTTPException:
+            log.exception("Failed to send free-text AskUserQuestion UI to thread")
+            return None
+
+        try:
+            return await view.wait_for_text()
+        finally:
+            for item in view.children:
+                setattr(item, "disabled", True)
+            try:
+                await msg.edit(view=view)
+            except discord.HTTPException:
+                log.debug("Could not disable free-text view (already gone)")
 
 
 # ---------------------------------------------------------------------------
@@ -303,9 +342,10 @@ class _AskSelect(discord.ui.Select):
         options: list[discord.SelectOption],
         min_values: int,
         max_values: int,
+        placeholder: str = "Pick an option…",
     ) -> None:
         super().__init__(
-            placeholder="Pick an option…",
+            placeholder=placeholder,
             min_values=min_values,
             max_values=max_values,
             options=options,
@@ -328,9 +368,211 @@ class _AskSelect(discord.ui.Select):
             log.debug("Failed to defer select interaction (already responded)")
 
 
+# ---------------------------------------------------------------------------
+# Paginated select view for > 25 options
+# ---------------------------------------------------------------------------
+
+
+class AskPaginatedSelectView(_BaseAskView):
+    """Paginated select menu for > 25 options.
+
+    Splits the option list into pages of ``page_size`` items, each rendered
+    as a Discord select menu.  Prev/Next buttons navigate between pages.
+    """
+
+    def __init__(
+        self,
+        labels: list[str],
+        descriptions: list[str],
+        *,
+        multi_select: bool,
+        timeout: float = DEFAULT_TIMEOUT_SECONDS,
+        page_size: int = _DISCORD_SELECT_MAX_OPTIONS,
+    ) -> None:
+        super().__init__(timeout=timeout)
+        self._labels = labels
+        self._descriptions = descriptions
+        self._multi_select = multi_select
+        self._page = 0
+        self._page_size = page_size
+        self._total_pages = max(1, (len(labels) + page_size - 1) // page_size)
+        self._build_page()
+
+    @property
+    def page(self) -> int:
+        return self._page
+
+    @property
+    def total_pages(self) -> int:
+        return self._total_pages
+
+    def _build_page(self) -> None:
+        self.clear_items()
+        start = self._page * self._page_size
+        end = min(start + self._page_size, len(self._labels))
+
+        options: list[discord.SelectOption] = []
+        for i in range(start, end):
+            desc = self._descriptions[i] if i < len(self._descriptions) else ""
+            options.append(
+                discord.SelectOption(
+                    label=self._labels[i][:_DISCORD_SELECT_LABEL_MAX] or f"Option {i + 1}",
+                    description=desc[:_DISCORD_SELECT_DESC_MAX] if desc else None,
+                    value=str(i),
+                )
+            )
+
+        max_values = len(options) if self._multi_select else 1
+        self.add_item(
+            _AskSelect(
+                options=options,
+                min_values=1,
+                max_values=max_values,
+                placeholder=f"Page {self._page + 1}/{self._total_pages}",
+            )
+        )
+
+        if self._page > 0:
+            self.add_item(_PageButton("◀️ Prev", -1, self))
+        if self._page < self._total_pages - 1:
+            self.add_item(_PageButton("Next ▶️", 1, self))
+
+
+class _PageButton(discord.ui.Button):
+    """Navigation button for :class:`AskPaginatedSelectView`."""
+
+    def __init__(
+        self, label: str, direction: int, paginated_view: AskPaginatedSelectView
+    ) -> None:
+        super().__init__(label=label, style=discord.ButtonStyle.secondary)
+        self._dir = direction
+        self._pview = paginated_view
+
+    async def callback(self, interaction: discord.Interaction) -> None:
+        self._pview._page += self._dir
+        self._pview._build_page()
+        try:
+            await interaction.response.edit_message(view=self._pview)
+        except discord.HTTPException:
+            log.debug("Failed to edit paginated view (already gone)")
+
+
+# ---------------------------------------------------------------------------
+# Modal for free-text (open-ended) questions
+# ---------------------------------------------------------------------------
+
+
+class AskTextModal(discord.ui.Modal):
+    """Modal dialog for free-text answers to open-ended questions."""
+
+    answer = discord.ui.TextInput(
+        label="Your answer",
+        style=discord.TextStyle.paragraph,
+        max_length=4000,
+    )
+
+    def __init__(self, question: str, header: str) -> None:
+        super().__init__(title=header[:45] or "Question")
+        self.answer.label = question[:45] or "Your answer"
+        self._text_future: asyncio.Future[str | None] | None = None
+        self._text_result: str | None = None
+        self._text_resolved: bool = False
+
+    async def on_submit(self, interaction: discord.Interaction) -> None:
+        self._text_resolved = True
+        self._text_result = self.answer.value
+        if self._text_future is not None and not self._text_future.done():
+            self._text_future.set_result(self.answer.value)
+        try:
+            await interaction.response.defer()
+        except discord.HTTPException:
+            pass
+
+    async def on_error(self, interaction: discord.Interaction, error: Exception) -> None:
+        log.exception("AskTextModal error")
+        self._text_resolved = True
+        self._text_result = None
+        if self._text_future is not None and not self._text_future.done():
+            self._text_future.set_result(None)
+
+
+class _ModalTriggerView(discord.ui.View):
+    """View with a single button that opens :class:`AskTextModal`.
+
+    Because Discord Modals can only be sent as an interaction response,
+    we render a button first and open the Modal when the user clicks it.
+    """
+
+    def __init__(
+        self,
+        question: str,
+        header: str,
+        timeout: float = DEFAULT_TIMEOUT_SECONDS,
+    ) -> None:
+        super().__init__(timeout=timeout)
+        self._question = question
+        self._header = header
+        self._text_future: asyncio.Future[str | None] | None = None
+        self._text_result: str | None = None
+        self._text_resolved: bool = False
+
+        btn = discord.ui.Button(
+            label="✏️ Type your answer",
+            style=discord.ButtonStyle.primary,
+            custom_id="clauded_ask_modal_trigger",
+        )
+        btn.callback = self._open_modal
+        self.add_item(btn)
+
+    async def _open_modal(self, interaction: discord.Interaction) -> None:
+        modal = AskTextModal(self._question, self._header)
+        # Wire the modal's future so on_submit resolves *our* waiter.
+        modal._text_future = asyncio.get_running_loop().create_future()
+
+        try:
+            await interaction.response.send_modal(modal)
+        except discord.HTTPException:
+            log.debug("Failed to send AskTextModal (already responded)")
+            self._set_text(None)
+            return
+
+        # Wait for modal submission
+        try:
+            text = await asyncio.wait_for(modal._text_future, timeout=self.timeout or 300)
+        except (asyncio.TimeoutError, asyncio.CancelledError):
+            text = None
+        self._set_text(text)
+        self.stop()
+
+    async def wait_for_text(self) -> str | None:
+        """Block until the user submits text or the view times out."""
+        if self._text_resolved:
+            return self._text_result
+        if self._text_future is None:
+            self._text_future = asyncio.get_running_loop().create_future()
+        try:
+            return await self._text_future
+        except asyncio.CancelledError:
+            return None
+
+    async def on_timeout(self) -> None:
+        self._set_text(None)
+
+    def _set_text(self, value: str | None) -> None:
+        if self._text_resolved:
+            return
+        self._text_resolved = True
+        self._text_result = value
+        fut = self._text_future
+        if fut is not None and not fut.done():
+            fut.set_result(value)
+
+
 __all__ = [
     "InteractionHandler",
     "AskButtonView",
     "AskSelectView",
+    "AskPaginatedSelectView",
+    "AskTextModal",
     "DEFAULT_TIMEOUT_SECONDS",
 ]
